day8: use records
Ok, lesson learned about the state monad, it's great but slow to
write, and when you need to change the state you have to go everywhere
in your code where you get/put it.  If the state is a record, combined
with NamedFieldPuns, it's basically imperative programming and you can
add extra state no problem.